requirements.txt
requests
packaging
beaker-py==0.2.6
click>=8.0,<8.1
click-help-colors>=0.9.1,<0.10
